Implement isArray as an attribute of other decorators.
Our `IsArray` decorator has turned out to be somewhat awkward because
it cannot practically specify type-specific validation of the array's
items. For example, it cannot say that an array contains integers,
all of which are greater than zero.

The best alternative is to add an `isArray` option to each typed
decorator. We'll likely deprecate `IsArray` eventually as well.