Enable @Validate as a common behavior
We want to enable the `@Validate` decorator for any decorator, so we refactor
the shared array decorators into a shared set of decorators of multiple types
that includes both the array and this validate logic.

As part of making this work, I also split the interfaces and options definitions
as the former is mostly internal and the latter is external. I also replaced
`Constructor` with the `@nestjs/common` `Type` declaration (which is equivalent).