Introduce DTO operators to pick/omit fields.
As business logic becomes more and more distributed across multiple services,
it is useful to have DTOs that are shared between services. However, we also
prefer that our services be able to define their own types and not always be
forced to use 100% of the the shared DTOs.

While *extending* an existing DTO is a simple matter of sub-classing the DTO
type, *narrowing* a DTO is more complicated because the metadata created by
DTO decorators persists in the class hierarchy, so a simple `Pick` or `Omit`
type operator is not sufficient.

This PR starts to introduce DTO-aware operators that interact with decorator
metadata, starting with OpenAPI data.

 -  Adds a `src/operators` tree and implements `pick/narrow`, stubbing out
    the transformer/validator parts.

 -  Adds a `src/testing` tree to help with DTO testing.

    While this code could be kept under `__tests__`, experience in downstream
    usages of this library suggest that there is value in reusing this test code.

 -  Upgrades several libraries for good measure.