Implement omit/pick for class-validator.
`class-validator` implements (and exposes!) `MetadataStorage` that records the list of
validators for each class along with the validated `propertyName`. The validator runtime
walks a validated object's class hierarchy within this metadata to determine the list of
applicable validators.

This PR extends (and completes) the `omit`/`pick` implementation by applying the `@IsOptional()`
decorator to the appropriate fields in the derived class, thereby allowing parent class
validators to be skipped for fields that are not present in the subclass.