Add experimental Union decorator.
We have at least one business case where a single field can contain either a numeric
value or a string value. While we _could_ implement this scenario using two fields,
it's a lot nicer (in TypeScript) to use a union type because it allows you express
that exactly one value is defined.

However, in the DTO layer, union types are trickier, perhaps so much that the benefit
of union types isn't worth the complexity:

 - You have to write a transformer layer that is aware of validation and essentially
   does a try/catch on each possible transformation contingent on validation.

 - You have to use `oneOf` at the OpenaPI layer, which, while possible, isn't clean.

This PR demonstrates how a `@Union()` decorator that implements `oneOf` might look
and hopefully demonstrates the consequences.