Automatically transform nullToUndefined
If a decorator is `optional` but is NOT `nullable`, we can automatically
convert null inputs to undefined.

In theory, this conversion could swallow some error conditions. We're basically
trading off the ability to catch (and fail upon) null being provided in case
where it's not allowed against the ability to omit conversion in cases where
null is more common (e.g. databases). This trade-off feels worth it.