Remove basic flavor.
We pretty much always want to use the `strict` flavor because it is more explicit. To the best
of our knowledge, we no longer use the `basic` flavor; it was introduced because some legacy
code balked at objects that produced keys (but undefined values) for omitted, optional properties.

We believe the legacy code was fixed.