Switch to dto-decorators for implementation.
The underlying implementation of `@hippo-oss/nest-dto` has become too complex and too difficult
to extend. We've now re-designed the implementation so that:

 - The basic decorator types and composition operations live in `@hippo-oss/dto-decorators`.
 - The `class-transformer` and `class-validator` flavors live in `@hippo-oss/class-decorators.`
 - The `@nestjs/swagger` implementation lives in `@hippo-oss/openapi-decorators`.

This rewrite makes this library mostly a shim around the other libraries, plus some utilities
for modifying (pick/omit) decorated types and for testing OpenAPI schemas.

This change also deprecated the `basic` flavor and the `@IsArray` decorator.

In exchange, we get a more maintainable code base, the ability to run a subset of the decorators
nicely (e.g. excluding `@nestjs/swagger` in React apps), and a much simpler process for defining
new decorator implementations.